fastapi_auth_project/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
# passlib[bcrypt]
python-jose
pydantic-settings
